backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
sqlalchemy==2.0.12
psycopg2-binary==2.9.6
alembic==1.10.0
pydantic==1.10.7
passlib[bcrypt]==1.7.4
bcrypt==3.1.7
python-dotenv==0.21.0
pytest==7.2.2
pytest-asyncio==0.19.0
docker==5.0.3
python-jose==3.3.0 
email-validator==1.3.1